+++ lib/cleanup.decorator.spec.ts
@@ -1,22 +1,28 @@
 import { Subject, Subscription } from 'rxjs';
-import { RxCleanup, RxCleanupGlobal } from './cleanup.decorator';
+import { RxCleanup } from './cleanup.decorator';
 import { DoneSubject } from './done-subject';
 
 class TestContext {
-  cleanedUp = false;
+  cleanedUp = 0;
   private readonly done$ = new DoneSubject();
-  cleanUp() {
-    this.cleanedUp = true;
+  ngOnDestroy() {
+    ++this.cleanedUp;
   }
 }
 
 class TestContext1 extends TestContext {
   readonly subject$ = new Subject();
   sub: Subscription | null = null;
+  ngOnDestroy() {
+    super.ngOnDestroy();
+  }
 }
 
 class TestContext2 extends TestContext {
   readonly otherSubject$ = new Subject();
+  ngOnDestroy() {
+    super.ngOnDestroy();
+  }
 }
 
 describe('RxCleanup', () => {
@@ -26,27 +32,26 @@ describe('RxCleanup', () => {
     instance1.sub = subj$.subscribe();
     const instance2 = new TestContext2();
 
-    RxCleanupGlobal.funcCleanUp = 'cleanUp';
     RxCleanup()(TestContext.prototype, 'done$');
     RxCleanup()(TestContext1.prototype, 'subject$');
     RxCleanup()(TestContext1.prototype, 'sub');
     RxCleanup()(TestContext2.prototype, 'otherSubject$');
 
-    expect(instance1.cleanedUp).toBe(false);
+    expect(instance1.cleanedUp).toBe(0);
     expect((instance1 as any).done$.isStopped).toBe(false);
     expect(instance1.subject$.isStopped).toBe(false);
     expect(instance1.sub.closed).toBe(false);
-    instance1.cleanUp();
-    expect(instance1.cleanedUp).toBe(true);
+    instance1.ngOnDestroy();
+    expect(instance1.cleanedUp).toBe(1);
     expect((instance1 as any).done$.isStopped).toBe(true);
     expect(instance1.subject$.isStopped).toBe(true);
     expect(instance1.sub.closed).toBe(true);
 
-    expect(instance2.cleanedUp).toBe(false);
+    expect(instance2.cleanedUp).toBe(0);
     expect((instance2 as any).done$.isStopped).toBe(false);
     expect(instance2.otherSubject$.isStopped).toBe(false);
-    instance2.cleanUp();
-    expect(instance2.cleanedUp).toBe(true);
+    instance2.ngOnDestroy();
+    expect(instance2.cleanedUp).toBe(1);
     expect((instance2 as any).done$.isStopped).toBe(true);
     expect(instance2.otherSubject$.isStopped).toBe(true);